fix: collapse current behavior into background+motivation (#216)
## Description:
We received feedback that some users felt like they were repeating
themselves when filling out details for the the "Current Behavior"
section and "Background and Motivation" sections. This PR collapses
those two together to make submitting FRs easier and smoother.

## Is this change user facing?
<!-- A user facing change is one that you should expect a day-to-day
user to encounter or if the change requires user-action upon or before
upgrading. If in doubt, select "Yes" -->
* [X] Yes
* [ ] No

<!-- If yes, please add the `user facing` label to this Pull Request -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->

Co-authored-by: Derek Lee <derek@Dereks-MacBook-Pro.local>