docs: hyperlink the *entire* calendly banner (#229)
## Description:
Right now, the link for our calendly link in the Docs is only
hyperlinked for the word "here". This PR proposes to hyperlink the
entire *text* in the banner in our docs.

## Is this change user facing?
<!-- A user facing change is one that you should expect a day-to-day
user to encounter or if the change requires user-action upon or before
upgrading. If in doubt, select "Yes" -->
* [X] Yes
* [ ] No

<!-- If yes, please add the `user facing` label to this Pull Request -->

## References (if applicable):
https://github.com/kurtosis-tech/kurtosis/pull/157