feat!: Reduce wait default timeout from 15 minutes to 10 seconds (#211)
## Description:

For command, 15 minute timeout was picked with the Ethereum module in
mind, but for most applications this seems excessive. This PR reduces
the waiting time from 15 minutes to 10 seconds, that seems to be a
reasonable timeout across requests and execs.

## Is this change user facing?
<!-- A user facing change is one that you should expect a day-to-day
user to encounter or if the change requires user-action upon or before
upgrading. If in doubt, select "Yes" -->
* [X] Yes
* [ ] No

<!-- If yes, please add the  label to this Pull Request -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->
Closes https://github.com/kurtosis-tech/kurtosis/issues/173