docs: update the headers for the delay dist objects (#252)
## Description:
Similar to #189, this PR breaks out the header titles for the docs about
the packet delay distribution objects

## Is this change user facing?
<!-- A user facing change is one that you should expect a day-to-day
user to encounter or if the change requires user-action upon or before
upgrading. If in doubt, select "Yes" -->
* [X] Yes
* [ ] No

<!-- If yes, please add the  label to this Pull Request -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->

Changelog picked up from commits here:

docs: update the headers for the delay dist objects

Co-authored-by: Derek Lee <derek@Dereks-MacBook-Pro.local>