fix: Lock default context in contexts config (#277)
## Description:
Disable possibility to add a new context with name `default` or removing
the `default` context. This can lead to unexpected situation where
someone ends up having a single remote context, and this remote context
not working properly on the server. This might be something we want to
allow in the future, but for now disabling it will save us some edge
cases pain.

## Is this change user facing?
<!-- A user facing change is one that you should expect a day-to-day
user to encounter or if the change requires user-action upon or before
upgrading. If in doubt, select "Yes" -->
NO

<!-- If yes, please add the `user facing` label to this Pull Request -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->