feat: Kurtosis backend can now connect to a remote Docker backend (#285)
## Description:
This PR adds a specific type of backend consisting of both the local
Docker backend and a remote Kurtosis backend (which is effectively
Docker). CLI and Engine are using this specific backend to be able to
offload some containers to the remote backend based on the current
context config

## Is this change user facing?
NO -- since we don't have the possibility to switch backend right now,
this is effectively not exposed
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->