docs: Improve runtime value typing documentation (#298)
## Description:
<!-- Describe this change, how it works, and the motivation behind it.
-->
Right now our `gojq` output is typed and is parsed into its `Starlark`
equivalent, but there is not mention to this behavior on the docs. Given
that several of our function calls (like `assert`, `wait`) requires
types to match, I'm updating the documentation with examples on how to
parse types.

## Is this change user facing?
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->
YES

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->