fix: Remove Quickstart, SDK, and CLI links from the navbar (#449)
## Description:
The Docusaurs navbar, very frustratingly, is NOT versioned with the
versioned docs. Therefore, the navbar must work across ALL versions of
the docs. If you want to rename something the navbar is pointing to, you
either have to:

1) not do it or
2) rename it in _historical_ versions of the docs (which is a big breach
of historical hermeticism)

It really seems like putting links in the navbar that point internal to
the docs is a bad idea, so I'm removing them for now.

## Is this change user facing?
YES

Co-authored-by: mieubrisse <mieubrisse@gmail.com>