feat: Raise file size limit to 100mb for file downloads and uploads (#542)
## Description:
This change raises the maximum file size limit to 100mb per file for
file uploads and downloads. This also applies to Starlark package (with
the entire package size compressed being under this limit). This limit
is purely arbitrary and is set to avoid dumping too much data into the
Kurtosis by mistake. It can be raised or removed if necessary.

## Is this change user facing?
YES
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->