fix: Remove the GRPC proxy port from the engine and from the APIC (#626)
## Description:
The engine container listens on the GRPC proxy port even if this port is
not used anymore. This PR removes the GRPC proxy port from the engine
and the code supporting it. We took the opportunity to also remove the
GRPC proxy port support from the APIC and the GRPC-WEB support since
they are not used anymore.

## Is this change user facing?
NO

## References (if applicable):
Closes #560

---------

Co-authored-by: leoporoli <leandroporoli@gmail.com>