docs: change --enclave-id to just --enclave (#629)
## Description:
When you do `kurtosis run --enclave $NAME_HERE`, you are essentially
telling Kurtosis to run a Starlark script against the enclave named
`$NAME_HERE`, or to create a new one with that name if one does not
exist.

In our docs, it right now says `--enclave-id` which is wrong. This PR
changes that

## Is this change user facing?
YES

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->

Co-authored-by: Victor Colombo <victorcolombo@users.noreply.github.com>