refactor: Set CGO_ENABLED=0 for APIC and Engine (#641)
## Description:
<!-- Describe this change, how it works, and the motivation behind it.
-->
We use to set `CGO_ENABLED=1` to allow the dynamic linking of the
Kubernetes plugin binary in runtime. Now that Kubernetes code has been
merged, we can revert back into shipping a cross compiled single binary
of both APIC/Engine that does not depend on a `libc`.

This also reduces:

- the setup burden on people developing Kurtosis
- image size

## Is this change user facing?
NO
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->