feat: Start and Stop service Starlark instructions for Docker (#694)
## Description:
Add `plan.start_service` and `plan.stop_service` Starlark instructions
for the Docker backend. Support for the K8S backend first requires an
update to the Gateway to manage the port forwarder connection during
stop and start operations. The k8s backend code itself is ready and the
changes (small) are in this PR.

No new backend methods are introduced, the goal being to re-use the
current code as much as possible. All the changes under `api/` and
`cli/` are related to the renaming of the API method `StartServices` to
`AddServices`. Most of the code added is unit and integration tests. The
map of service registrations at the service network layer is used as
part of the existing start service process so I am using this map to
store the service status and config required for the restart process. I
thought about using a map of created `Services` but it is not as clean
as leveraging the map of service registrations.

Added some PR comments to help with the review.

Follow-up work in subsequent PRs:

- Update gateway to manage the port forwarder connection during stop and
start operations and enable support for k8s.
- Update documentation.

## Is this change user facing?
YES

## References (if applicable):
#673