feat: added minimum resources for containers on k8s to the docs. (#708)
Updated the docs for:

Users can add min resource requirements for their containers on
Kubernetes and update the doc accordingly. Along with that, we are
replacing `cpu_allocation` and `memory_allocation` with `max_cpu` and
`max_memory` fields in ServiceConfig.

We will deprecate `cpu_allocation` and `memory_allocation` in two weeks
from now. An example to set resources:

```
ServiceConifg( 
..., 
 max_cpu = 1000,
 min_cpu = 500,
 max_memory = 1024,
 min_memory = 512,
 )
```