feat: Stop and start service support in the CLI (#767)
## Description:
It is currently possible to stop and restart a service using the
Starlark `plan.stop_service` and `plan.start_service` instructions. This
change adds the same support to the CLI with the command `service stop`
and `service start`. Both commands take as input an enclave identifier
and a service identifier similar to the `service logs` and `service rm`
commands. This change also updates the docs CLI section and adds two
basic CLI tests to CI.

## Is this change user facing?
YES

## References (if applicable):
Closes #705