feat: Implement idempotent run v0 (#769)
## Description:
This feature saves the current state of the enclave after each starlark
run and tries to optimize the next execution to execute only the
instructions that changed compare to previous run. For this V0, the
optimisation is "append-only". I.e. the resolution of instructions will
succeed only if what changed is at the end of the instruction queue

## Is this change user facing?
NO
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->