feat: Start engine remotely with remote backend config when the context is remote (#963)
## Description:
This is the first part to [support the bastion with the docker and k8s
remote
backend](https://www.notion.so/kurtosistech/Kloud-Bastion-Docker-and-K8S-backends-71a19e630ed0454a984042e1cafb1a88?pvs=4#8f3b10e02c874bdd9b9aad22acc76e14).
The CLI starts the engine remotely when the context is remote. The
engine reads the remote backend config at startup. The CLI requests the
portal client to forward the engine port locally. The remote backend
config supplier and the remote context backend are removed.

Manually tested with local and remote context against a Kloud instance
acting as the bastion.

The deployment tool will be updated in a subsequent PR.

## Is this change user facing?
YES