feat: Persistent directories for Kubernetes (#1036)
## Description:
Adds the docker backend to support persistent directories for user
services. For now, we use `hostPath` persistent volumes, which will
break if the user service is restarted onto a different k8s node than
what it was before. There's plenty of ways to fix that (using k8s
`local` persistent volumes, re-scheduling the pod on the same node as
where it was before, using a shared block store like EBS on AWS, etc)
but we are lacking concrete use cases to choose the best option right
now.
`hostPath` works for now and is the simplest implementation.

## Is this change user facing?
NO
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->