docs: add short term roadmap to our docs (#1037)
## Description:
We've fielded a lot of questions about what features/changes are coming
down the pipeline. This small PR adds a new page to our docs that shares
our short-term roadmap with users. The time horizon used is 1 quarter (3
months) and we will constantly update this page as things change.

## Is this change user facing?
YES
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->