feat: Update files if already present in enclave (#1066)
Right now, when a Starlark package creates a file artifact (via
`upload_files`, `render_template`, or `store_service_file`), if a file
artifact already exist with that name inside the enclave, it fails. This
is a big pain for idempotency.
This PR introduces files artifact updates. From now on, when a files
artifact is "created" via starlark and already exists in the enclave
(based on its artifact name), Kurtosis will check the equality of the
files based on their hash, and if they are different the files artifact
will be replaced by the new one (and of course it no-ops if they are the
same).
No change requires in the Starlark packages

YES
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->