docs: update service delete to service rm (#1093)
## Description:
Previously, the cli reference for `service rm` was titled `service
delete`. This PR changes the title to `service rm` to match our page on
`enclave rm` and also because the actual command is `kurtosis service
rm` and not `kurtosis service delete`, so this change ensure that the
reference is consistent with the actual command.

## Is this change user facing?
YES

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->