feat: Portal automatic start and stop on context change (#1086)
## Description:
The user currently needs to make sure that the portal is running before
switching to a remote context. This change does that automatically for
the user. This change also stops the portal when the context switches
back to local.

Add the portal release tag to install so we can pin a portal version
instead of using the latest. The portal release tag can also be set to
"latest". This PR pins the portal version to 0.0.4

Upcoming changes:
- Release portal 0.0.5 containing the support for remote endpoints.
- Update the mono-repo to pass the remote endpoint type to the portal
client forward port calls. Set the portal version to latest.

## Is this change user facing?
YES

## References (if applicable):
Closes #970