fix: Forward the engine port after verifying that an engine container is running and before initializing the engine client (#1099)
## Description:
The engine port is forwarded at the start of each low level kurtosis
command. However, it is not enough for all CLI operations such as
context switching where the CLI switches engine.

`GetEngineStatus` is on multiple CLI paths and it first checks if an
engine container is running before getting the engine info via the
engine port. This change forwards the engine port before the engine
client is initialized. We also remove the engine port forwarding from
the `engine status` command since it calls `GetEngineStatus`.

## Is this change user facing?
NO