feat!: Portal remote endpoints and forward port wait until ready option (#1124)
## Description:
This change is for the Portal release 0.0.5 - remote endpoints
(https://github.com/kurtosis-tech/kurtosis-portal/pull/44) and 0.0.6 -
forward port wait until ready option
(https://github.com/kurtosis-tech/kurtosis-portal/pull/46). The remote
endpoint type and the wait until ready flag are passed to the forward
port calls. Both changes are in this single PR to minimize the Portal
server upgrades.

Those are internal changes but they break the CLI if the user Kloud
instance is not running Portal 0.0.6.

Portal 0.0.5 has been already released but is not deployed. The next
step is to release 0.0.6 and then merge this PR.

## Is this change user facing?
YES
The user Kloud instance requires Portal 0.0.6.

## References (if applicable):
https://github.com/kurtosis-tech/kurtosis-portal/pull/44
https://github.com/kurtosis-tech/kurtosis-portal/pull/46