chore: update build files (#1169)
## Description:
There was a minor debugging logging change that didn't make it into the
build files of [this
PR](https://github.com/kurtosis-tech/kurtosis/pull/1166) (i.e. the code
change made it. but the build wasn't updated to include it). I am
rebuilding it here to include the latest.

## Is this change user facing?
NO
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->