fix: Point to the engine restart command as part of the context switch failure remediation to not conflict with lower level commands (#1191)
## Description:
The engine remediation command in the context switch logic sometimes
conflicts with lower level engine remediation commands. One example is
when the running engine version does not match the CLI version, the user
is presented with two remediation commands: `engine start` and `engine
restart`. Only the second one remediates the issue. This change updates
the engine remediation command in the context switch logic to be `engine
restart` which also works if the engine is not running.

Note that we don't want to restart the engine automatically when there
is a version mismatch because different users could be using different
CLI versions and we don't want to end up in a tug of war match.

## Is this change user facing?
YES