docs: add caveat of 200 line truncation in kurtosis service logs (#1209)
## Description:
A user noticed that `kurtosis service logs` only displays 200 lines of
logs in the terminal. This was an arbitrary number chosen for UX. This
PR adds a tip in the docs for `kurtosis service shell` to recommend
users use the `kurtosis dump` command for a full snapshot of logs or the
`-f` flag for a stream of logs (i.e. follow the logs).

## Is this change user facing?
YES

## References (if applicable):
https://github.com/kurtosis-tech/kurtosis/issues/1188