fix: Remove 'kurtosis version' depending on the engine (#1207)
## Description:
I was surprised that `kurtosis version` was taking 7 seconds without any
extra output, and so I checked out the code to discover that the
`kurtosis version` command was depending on the engine. This felt wrong
to me, given that we already have `engine status` to display the version
(and it's more detailed besides - e.g. if the engine is unreachable, you
get a message saying why). If we want one command to display the status
of both CLI and engine(which would also be fine), we should consolidate
the two so that we're not duplicating partial functionality of `engine
status` in `version`.

All that said, I'd also be fine if we went the other way and built out
`version` heavily to properly communicate what's happening with the
engine, and why it's hanging for 7s. This seemed like it would be a
bigger change though and my Goland is already yelling at me for Go
workspaces stuff that I don't know how to fix, so I went with the
quickest change I could.

## Is this change user facing?
YES

## References (if applicable):
Fixes #1206

Co-authored-by: mieubrisse <mieubrisse@gmail.com>