docs: add extra mev-flood caveat for failed transactions in logs (#1208)
## Description:
A user noticed that there were failed transactions from mev-flood when
using the eth2-package in `full-mev` mode. After debugging with the user
and some information gathering from the flashbots team, we discovered
that this behavior is normal due to the speed at which txns are sent to
the network from mev-flood.

This PR adds a caveat in the guide to inform future users of this
behavior.

## Is this change user facing?
YES


## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->