feat: always pull latest image (#1267)
## Description:
Prior to this PR, Kurtosis would use the local image if it was
available, and if it wasn't, pulled the image from Dockerhub (or other
container registry). This PR reverses that behavior by always attempting
to pull the latest image, and using the local image, if available, only
if the pull fails. For users, this means they will always be using the
latest version of the images they're working with.

Through testing, I've determined the impact on speed is negligible but
I've kept a `FetchLocalImage` function with the old behavior in case we
need to rollback this change. I've also removed a redundant `FetchImage`
from the `StartServices` as it is already being called for each service
during validation phase and in `CreateAndStartContainerArgs`.

## Is this change user facing?
NO

## References
https://github.com/kurtosis-tech/kurtosis/issues/1103