fix: warn instead of failing for json log line parse error (#1336)
## Description:
Occassionally, we hit a race condition between the `LogsAggregator`
writing a log line while `service logs` is reading and streaming log
files. This results in a half written json log line being read while
streaming logs and a parsing error to occur. Instead of failing
immediately, this PR logs a `Warn` instead and continues to read the log
file.


## Is this change user facing?
YES