fix: restart log aggregator on failure (#1371)
## Description:
This change adjusts the log aggregator container configuration so that
it's Docker always restarts the container when it detects it has exited
with non-zero exit code. It improve how we handle errors related to log
aggregator not existing when creating an enclave.

This is a bandaid that should address
https://github.com/kurtosis-tech/kurtosis/issues/1311 but doesn't
address the issue entirely. More context in issue's discussion.

## Is this change user facing?
YES (users should not experience
https://github.com/kurtosis-tech/kurtosis/issues/1311 anymore)

## References (if applicable):
https://github.com/kurtosis-tech/kurtosis/issues/1311