fix: make vector use ISO week time (#1497)
## Description:
There was a bug where the `LogsAggregator` was using a different time
format for week than the `LogsDatabaseClient`.

In special cases (like today), the `LogsAggregator` and
`LogsDatabaseClient` disagreed about what week it was. The former said
it was week 41 and the latter said it was week 40. This disagreement
prevented users from being able to pull logs today.

This change makes them agree from now on by making them use the same
week format : )

## Is this change user facing?
YES (no more bug)