feat: rename kurtosis context "switch" to "set" (#1537)
## Description:
Changes the command `kurtosis context switch` to `kurtosis context set`,
to match the similar command `kurtosis cluster set`.

I've changed the cli command itself, plus the context config store.
There's also the portal manager, which I've left as switch for now, as
it extends into the kurtosis-portal project; I think we'll also need to
do that?

We're missing documentation on contexts and this cli command. I don't
know what to write for it yet so I've added
https://github.com/kurtosis-tech/kurtosis/issues/1536

## Is this change user facing?
YES
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
https://github.com/kurtosis-tech/kurtosis/issues/1508

---------

Co-authored-by: Omar <omar@omar>