feat: Add create enclave utils to SDK (#1550)
## Description:
<!-- Describe this change, how it works, and the motivation behind it.
-->
A very idiomatic pattern in Go is to use `defer` and closer functions to
cleanup resources, instead of relying on memory to call the correct SDK
calls when the test/setup is done. This is very common on our test suit
that we have a util, that now is exposed to an external user.

This PR also introduces a single SDK call to get from starlark to
enclave, something that was not possible before.

## Is this change user facing?
YES
<!-- If yes, please add the "user facing" label to the PR -->
<!-- If yes, don't forget to include docs changes where relevant -->

## References (if applicable):
<!-- Add relevant Github Issues, Discord threads, or other helpful
information. -->
https://github.com/kurtosis-tech/kurtosis/issues/1152