fix: handle error and fix rendering bug (#1617)
## Description:
Added logic to handle error when script fails to run and returns an
error. Before the status of the execution would just hang as being in
progress, but now it updates the status to be an error (when the error
is returned from the EM API).

As a followup I've created issue #1618 to provide better error reporting
in the UI.

## Is this change user facing?
YES

## References (if applicable):
Closes: https://github.com/kurtosis-tech/kurtosis/issues/1587