chore: Simplify enclave-manager-sdk typescript dependencies (#1595)
## Description:

The enclave manager sdk has a large number of (I think) unnecessary
dependencies, this PR removes a lot of them and tidies up package.json.

Additionally, specify `yarn --frozen-lockfile` in the
`enclave-manager/api/typescript/scripts/build.sh` so that builds will
fail if dependencies are not committed properly.

This change removes the dependency in the enclave manager from the
general typescript `kurtosis-sdk` - I'm not sure why this dependency is
present as it's not required by any code generated by the
enclave-manager-sdk.

## Is this change user facing?
NO

## References (if applicable):
N/A