ci: increase time to persist logs (#1621)
## Description:
This is a hack to prevent search logs ci tests from failing
intermittently as much. The underlying issue is that the time it takes
for logs aggregator to persist logs is nondeterministic and
occassionally, takes longer on some ci runs for some reason. I've
created an issue to track this.

## Is this change user facing?
NO

## References:
https://github.com/kurtosis-tech/kurtosis/issues/1620