feat: preload package in new enclave manager (#1657)
## Description:
This PR implements the 'preload package' functionality in the new
enclave manager. The package arguments are base64 encoded json
representations of the configure form values - this means that invalid
configuration can be copied (perhaps useful for debugging?)

Additionally this PR refactors the `EnclaveFullInfo` object to contain
`true-myth` `Result` objects for the services, artifacts and starlark
run properties. The error type for these Result objects is simplified to
a `string` - I thought originally that having the actual Error object
might be useful (hence `MessageAndError`, but I've not used the `error`
property at all - instead just passing `message` to alerts.

### Demo


https://github.com/kurtosis-tech/kurtosis/assets/4419574/d068a8e9-fe0f-49ba-8eb4-e5e75545f268

## Is this change user facing?
NO (the new enclave manager is not deployed anywhere

## References (if applicable):

---------

Co-authored-by: Anders Schwartz <anders.schwartz@kurtosistech.com>