feat: enclave manager design tweaks and enclave view (#1639)
## Description:
This PR primarily adds support for the new enclave view to the enclave
manager, and picks up changes to the design of components. Significant
changes:
* The enclave view is added
* Theming has been updated to match the mocks. Now that I've taken a few
look over the mocks and am more familiar with which components are
special and which are common, where possible I've shifted to overriding
existing chakra styles rather than defining new ones for the brand
* The breadcrumb logic will now be used across the application, rather
than custom for enclaves (catalog etc)
* React router is used properly to load data, rather than loading in
effects. In order to provide the `kurtosisClient` to react router -
which can only be provided after initial render - the
`createBrowserRouter` is moved around a little to live inside the new
`KurtosisRouteProvider` instead.

### Feedback required:

* The mocks include container started times. I think the api needs to
have this added
* The mocks include file descriptions - is this actually stored? How can
it be obtained
* The port summary overview is not in the mocks - my implementation is
likely poorly designed.
* Is there an empty state for tables that can be used?

### Demo


https://github.com/kurtosis-tech/kurtosis/assets/4419574/192240bd-6cbf-4a69-9c7f-072783fd965e

## Is this change user facing?
NO (This is not yet in the deploy process)

## References (if applicable):
This PR is based on top of
https://github.com/kurtosis-tech/kurtosis/pull/1603 which should be
merged first.

---------

Co-authored-by: Anders Schwartz <anders.schwartz@kurtosistech.com>