feat: initial new enclave manager (#1603)
## Description:
This PR contains the beginnings of the new enclave manager ui.
Specifically it adds a new project in `enclave-manager/web`. Notable
inclusions:
* Continuing to use Chakra as a component libary, the
KurtosisThemeProvider injects components themed based on the mocks.
* The enclave manager sdk cllient is refactored into authenticated
(remote) and local versions, and provided to react components with
`useKurtosisClient` by the `KurtosisClientProvider`. This provider
blocks loading the application (with a spinner) until it's configured
the client to use.
* react router is used to compose the application.
* react-table is used to create a (reusable) table for showing the
enclaves list.

### Demo (WIP)


https://github.com/kurtosis-tech/kurtosis/assets/4419574/c026a1d9-313d-4e73-944a-2814399e039a

## Is this change user facing?
NO (this is not yet part of the build/deploy process)

## References (if applicable):
This PR is rebased on top of
https://github.com/kurtosis-tech/kurtosis/pull/1595 which should be
merged first.

---------

Co-authored-by: Anders Schwartz <anders.schwartz@kurtosistech.com>
Co-authored-by: Anders Schwartz <adschwartz@users.noreply.github.com>