feat: new em ui enclave logs (#1696)
## Description:
This PR implements the views for showing the logs in the new enclave
manager ui.

This change adds react-virtuoso for buffering the logs in a new
`LogViewer.tsx` component. `lodash` is used to `throttle` the changes -
I believe this solves the issue reported previously with log tailing
failing. Once the service log ui is implemented I'll validate this with
the log spamming package.

A utility `FeatureNotImplemented` modal is added to display more tidy
incomplete ends of exploration (per the mocks).

### Demo


https://github.com/kurtosis-tech/kurtosis/assets/4419574/18782856-5f0a-4e31-835d-6d840bbfda5c

## Is this change user facing?
NO - this is not currently deployed