fix: emui colormodefixer (#1716)
## Description:
The old emui set local storage `chakra-ui-color-mode` to `light`, but
the new emui is designed to work only in `dark` mode. This means that
users transitioning from the old emui to the new one see incorrectly
styled ui. This PR adds a component `ColorModeFixer` to the theme
provider which corrects the color mode to always be dark mode.

This PR additionally runs prettier:fix over the codebase to correct a
unnecessary formatting diff that was merged. I'll follow up this PR with
a simple check to make sure the code is formatted correctly pre merge.

### Screenshot of bad behaviour

![image](https://github.com/kurtosis-tech/kurtosis/assets/4419574/4ae6f900-9f55-4ad5-b7a8-559c2716043b)

## Is this change user facing?
YES

## References (if applicable):
* Raised on slack by @adschwartz and @leeederek