feat: new emui use monaco as json editor (#1733)
## Description:
This PR switches to using the monaco editor for json input. I've changed
the layout functionality to effectively match the previous
`CodeEditor.js` - which seems to work for me, though I can't explain why
calling `layout` twice (or so frequently) is necessary.

This PR also fixes the use of the Inter font

Additionally this PR fixes several pieces of feedback from Tise:
* breadcrumb styling
* Uppercase text in tags
* Tab hover
* `x selected` hover interactivity
* Update progress widget and colour used
* Implement the copy toast
* Fix table header icons and hover appearance.

### Demo


https://github.com/kurtosis-tech/kurtosis/assets/4419574/f01efc78-2022-4edf-a18d-01587f053390

## Is this change user facing?
YES