fix: emui optimistic data loading (#1771)
## Description:
This PR removes the use of react-router data loaders and actions,
instead managing this behaviour inside a new `EMUIAppContext`. This
allows us to implement optimistic ui patterns - letting us load the
initial enclave list as data become available, and updating it without a
backend fetch (on create/delete operations).

Also this PR fixes the issue identified with boolean values not being
persisted - in fact the values were read and persisted, just not
rendered correctly on subsequent load.

### Demo

As this is quite a substantial change, this demo quickly goes over the
majority of the data interactions in the emui.


https://github.com/kurtosis-tech/kurtosis/assets/4419574/0a7565bd-4e3c-43f4-bc10-8cb6b71eb3a3

## Is this change user facing?
YES